moved data to summaries and removed unecessary kruft. also fixed mempool height to 0 when raw transaction height is 0
